Rename integration tests to be about `GradleLifecycle`
Instead of `IsolatedAction` which is more of an implementation detail.
